Use a base template, prepare to provide the component to use on elements dynamically
